Move linter implementation classes into their own package
