+++ silence_lint_error/silence_lint_error.py
@@ -1,10 +1,6 @@
 from __future__ import annotations
 
-import json
-import re
 import subprocess
-import sys
-from collections import defaultdict
 from collections.abc import Sequence
 from typing import NamedTuple
 from typing import Protocol
@@ -12,7 +8,6 @@ from typing import TYPE_CHECKING
 
 import attrs
 
-from . import comments
 
 if TYPE_CHECKING:
     from typing import TypeAlias
@@ -37,230 +32,23 @@ class Linter(Protocol):
     def find_violations(
         self, rule_name: RuleName, filenames: Sequence[FileName],
     ) -> dict[FileName, list[Violation]]:
-        ...
+        """Find violations of a rule.
+
+        Returns:
+            Mapping of file path to the violations found in that file.
+
+        Raises:
+            ErrorRunningTool: There was an error whilst running the linter.
+        """
 
     def silence_violations(
         self, src: str, violations: Sequence[Violation],
     ) -> str:
-        ...
+        """Modify module source to silence violations.
 
-
-class Fixit:
-    name = 'fixit'
-
-    def __init__(self) -> None:
-        self.error_line_re = re.compile(r'^.*?@\d+:\d+ ')
-
-    def find_violations(
-        self, rule_name: RuleName, filenames: Sequence[FileName],
-    ) -> dict[FileName, list[Violation]]:
-        proc = subprocess.run(
-            (
-                sys.executable, '-mfixit',
-                '--rules', rule_name,
-                'lint', *filenames,
-            ),
-            capture_output=True,
-            text=True,
-        )
-
-        if proc.returncode and proc.stderr.endswith('No module named fixit\n'):
-            raise ErrorRunningTool(proc)
-
-        # extract filenames and line numbers
-        results: dict[str, list[Violation]] = defaultdict(list)
-        for line in proc.stdout.splitlines():
-            found_error = self._parse_output_line(line)
-            if found_error:
-                filename, violation = found_error
-                results[filename].append(violation)
-            else:  # pragma: no cover
-                pass
-
-        return results
-
-    def _parse_output_line(
-            self, line: str,
-    ) -> tuple[FileName, Violation] | None:
-        if not self.error_line_re.match(line):
-            return None
-
-        location, violated_rule_name, *__ = line.split(maxsplit=2)
-        filename, position = location.split('@', maxsplit=1)
-        lineno, *__ = position.split(':', maxsplit=1)
-
-        rule_name_ = violated_rule_name.removesuffix(':')
-        return filename, Violation(rule_name_, int(lineno))
-
-    def silence_violations(
-        self, src: str, violations: Sequence[Violation],
-    ) -> str:
-        [rule_name] = {violation.rule_name for violation in violations}
-        linenos_to_silence = {violation.lineno for violation in violations}
-
-        lines = src.splitlines(keepends=True)
-
-        new_lines = []
-        for current_lineno, line in enumerate(lines, start=1):
-            if current_lineno in linenos_to_silence:
-                leading_ws = line.removesuffix(line.lstrip())
-                new_lines.append(f'{leading_ws}# lint-fixme: {rule_name}\n')
-            new_lines.append(line)
-
-        return ''.join(new_lines)
-
-
-class FixitInline(Fixit):
-    """An alternative `fixit` implementation that adds `lint-fixme` comment inline.
-
-    This is sometimes necessary because `fixit` does not always respect `lint-fixme`
-    comments when they are on the line above the line causing the error. This is a
-    known bug and is reported in https://github.com/Instagram/Fixit/issues/405.
-
-    In some of these cases, placing the comment on the same line as the error can
-    ensure it is respected (e.g. for decorators).
-    """
-
-    def silence_violations(
-        self, src: str, violations: Sequence[Violation],
-    ) -> str:
-        [rule_name] = {violation.rule_name for violation in violations}
-        linenos_to_silence = {violation.lineno for violation in violations}
-        return comments.add_error_silencing_comments(
-            src, linenos_to_silence,
-            'lint-fixme', rule_name,
-        )
-
-
-class Flake8:
-    name = 'flake8'
-
-    def find_violations(
-        self, rule_name: RuleName, filenames: Sequence[FileName],
-    ) -> dict[FileName, list[Violation]]:
-        proc = subprocess.run(
-            (
-                sys.executable, '-mflake8',
-                '--select', rule_name,
-                '--format', '%(path)s %(row)s',
-                *filenames,
-            ),
-            capture_output=True,
-            text=True,
-        )
-
-        if proc.returncode and proc.stderr.endswith('No module named flake8\n'):
-            raise ErrorRunningTool(proc)
-
-        # extract filenames and line numbers
-        results: dict[FileName, list[Violation]] = defaultdict(list)
-        for line in proc.stdout.splitlines():
-            filename_, lineno_ = line.rsplit(maxsplit=1)
-            results[filename_].append(Violation(rule_name, int(lineno_)))
-
-        return results
-
-    def silence_violations(
-        self, src: str, violations: Sequence[Violation],
-    ) -> str:
-        [rule_name] = {violation.rule_name for violation in violations}
-        linenos_to_silence = {violation.lineno for violation in violations}
-        return comments.add_noqa_comments(src, linenos_to_silence, rule_name)
-
-
-class Ruff:
-    name = 'ruff'
-
-    def find_violations(
-        self, rule_name: RuleName, filenames: Sequence[FileName],
-    ) -> dict[FileName, list[Violation]]:
-        proc = subprocess.run(
-            (
-                sys.executable, '-mruff',
-                '--select', rule_name,
-                '--output-format', 'json',
-                *filenames,
-            ),
-            capture_output=True,
-            text=True,
-        )
-
-        if proc.returncode and proc.stderr.endswith('No module named ruff\n'):
-            raise ErrorRunningTool(proc)
-
-        # extract filenames and line numbers
-        all_violations = json.loads(proc.stdout)
-        results: dict[FileName, list[Violation]] = defaultdict(list)
-        for violation in all_violations:
-            results[violation['filename']].append(
-                Violation(
-                    rule_name=violation['code'],
-                    lineno=violation['location']['row'],
-                ),
-            )
-
-        return results
-
-    def silence_violations(
-        self, src: str, violations: Sequence[Violation],
-    ) -> str:
-        [rule_name] = {violation.rule_name for violation in violations}
-        linenos_to_silence = {violation.lineno for violation in violations}
-        return comments.add_noqa_comments(src, linenos_to_silence, rule_name)
-
-
-class Semgrep:
-    name = 'semgrep'
-
-    def find_violations(
-        self, rule_name: RuleName, filenames: Sequence[FileName],
-    ) -> dict[FileName, list[Violation]]:
-        proc = subprocess.run(
-            (
-                'semgrep', 'scan',
-                '--metrics=off', '--oss-only',
-                '--json',
-                *filenames,
-            ),
-            capture_output=True,
-            text=True,
-        )
-
-        if proc.returncode:
-            raise ErrorRunningTool(proc)
-
-        # extract filenames and line numbers
-        results: dict[FileName, list[Violation]] = defaultdict(list)
-        data = json.loads(proc.stdout)
-        for result in data['results']:
-            if result['check_id'] != rule_name:
-                continue
-
-            results[result['path']].append(
-                Violation(
-                    rule_name=result['check_id'],
-                    lineno=result['start']['line'],
-                ),
-            )
-
-        return dict(results)
-
-    def silence_violations(
-        self, src: str, violations: Sequence[Violation],
-    ) -> str:
-        [rule_name] = {violation.rule_name for violation in violations}
-        linenos_to_silence = {violation.lineno for violation in violations}
-
-        lines = src.splitlines(keepends=True)
-
-        new_lines = []
-        for current_lineno, line in enumerate(lines, start=1):
-            if current_lineno in linenos_to_silence:
-                leading_ws = line.removesuffix(line.lstrip())
-                new_lines.append(f'{leading_ws}# nosemgrep: {rule_name}\n')
-            new_lines.append(line)
-
-        return ''.join(new_lines)
+        Returns:
+            Modified `src` with comments that silence the `violations`.
+        """
 
 
 class NoViolationsFound(Exception):
